Fix disk cleanup bug

+++ stages/run_opensfm.py
@@ -43,10 +43,11 @@ class ODMOpenSfMStage(types.ODM_Stage):
             if args.optimize_disk_space:
                 for folder in ["features", "matches", "exif", "reports"]:
                     folder_path = octx.path(folder)
-                    if os.path.islink(folder_path):
-                        os.unlink(folder_path)
-                    else:
-                        shutil.rmtree(folder_path)
+                    if os.path.exists(folder_path):
+                        if os.path.islink(folder_path):
+                            os.unlink(folder_path)
+                        else:
+                            shutil.rmtree(folder_path)
 
         # If we find a special flag file for split/merge we stop right here
         if os.path.exists(octx.path("split_merge_stop_at_reconstruction.txt")):
